+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 I will try to keep track of changes between releases here
 
+- select_distro uses exit 123 for select_distro_prepare if chrooted
+- set_new_etc_profile - sets PATH if deploy script is aborted/completed
 - uses installed /etc/skel when creating accounts instead of copying from /opt/AOK
 - setup_final_tasks.sh now defines a full PATH including /usr/local/bin
 - Now hostname workaround can be enabled/disabled by running /usr/local/bin/aok